complete seeds & work on connections
